requirements.txt
asgiref==3.2.10
Django==3.1.2
django-filter==2.4.0
django-stubs==1.6.0
djangorestframework==3.12.1
djangorestframework-stubs==1.2.0
mypy==0.782
mypy-extensions==0.4.3
psycopg2==2.8.6
Pygments==2.7.1
pytz==2020.1
sqlparse==0.4.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
